🐛 Fix : Suspense 감싸기
